ADC reading working
For this, I had to use the ADC/Adc.h file that Tiago made it. The main.cpp function into the main folder has the call
ADC pot1(AN0); which stands for an instance of the object ADC giving the AN0 pin from the beagle bone. Tiago made an
enumartor to help it.